fix(sandbox): scope text direction to preview frames
