+++ apps/sandbox/app/shell/preview.tsx
@@ -4,12 +4,12 @@ import { Label } from '@app/components/ui/label';
 import { ResizableHandle, ResizablePanel, ResizablePanelGroup } from '@app/components/ui/resizable';
 import { Switch } from '@app/components/ui/switch';
 import { ToggleGroup, ToggleGroupItem } from '@app/components/ui/toggle-group';
+import type { ColorScheme, PreloadValue, TextDirection } from '@app/constants';
 import { LAYOUT_LABELS } from '@app/labels';
 import type { MediaId } from '@app/media';
 import type { CaptionsMode } from '@app/shared/captions';
 import type { SandboxLocaleTag } from '@app/shared/i18n/locale-meta';
 import type { AspectRatio } from '@app/shared/player-frame';
-import type { ColorScheme, PreloadValue, TextDirection } from '@app/shared/sandbox-listener';
 import type { SourceId } from '@app/shared/sources';
 import { ArrowTopRightOnSquareIcon } from '@heroicons/react/16/solid';
 import { observeResize } from '@videojs/utils/dom';
