--- packages/skins/src/gaps.md
@@ -1,5 +1,0 @@
-# VJSC skin gaps
-
-This file tracks known parity and anatomy work carried forward from the retired legacy skins. Remove an entry once it is implemented and verified for every affected skin, target, and styling output.
-
-No outstanding tracked parity gaps.
